Fixing typos and missing dependency.
I just followed the instructions provided in this tutorial.  I needed to make these changes to actually deploy the script. 
First, I made sure to require "@nomiclabs/hardhat-ethers", so that I would not get the errory TypeError: no propery found 'getContractFactory'. 
Second, I changed the 'k' in 'privatekey' to a capital to match it's later usage.  
Finally, I changed the quotes around the solidity version so that they match the other quotes.